# improve read.ctd.sbe preformance by reading directly from data file