This addresses a performance issue for large files identified in Issue #915. The change is to simply push back lines read using `readLines` to the connection, and then call `read.table()` on the connection directly. For whatever reason, using `read.table(text=...)` suffers a severe bottleneck in the `textConnection` code.