@dankelley 2016-04-07
It would be nice to document the speedup (I think in the issue, not here, since issues may be where people look).