@dankelley 2016-04-07
Thanks -- I'm going to look at the idea of reading just `n` lines at the start, to find the header.