@dankelley 2016-04-07
I tried a two-pass method, first reading 1000 lines to find the header and then reading the whole file. With the 2016-01-0063.cnv test file, this saved 0.3s to 0.4s (tests on both SSD and spinning-disk machines). This is a 10% speedup on a 14MB file since the total read times on the two test machines were about 2 and 3s.  I guess the speedup would be 1s on a 30MB (i.e. quite large) file. But I imagine that a 30MB file would also take 4 to 6s, so I don't think a 1s improvement will be too noticeable.

The downside to the two-pass is that the code is harder to read, and therefore it will be harder to maintain.

Therefore, I propose to drop the idea, on the grounds of code complexity and lack of dramatic speedup.

The **real** advance for this issue is clearly Clark's `pushback` call.  I wonder if we can get a statement here about the speedup that this yielded, Clark?