@dankelley 2016-04-07
Sorry, I see in https://github.com/dankelley/oce/issues/915 that the `pushback` speed up is 11X, from 11.2s to 0.9s for a 4.9MB file. If there is simple scale-up, then the 14MB file I tested would have taken about 35s in the old system, and about 3s in the new system. This is consistent with my timing. In other words: (a) avoiding `pushback` speeds up processing by 12X, while (b) a two-pass to read the header might yield a second 1.1X factor in speed. Given the issue of code complexity, I'll drop this.

Oh my, I see I'm writing this in the pull item not the issues item. I'm lost!