@richardsc 2016-04-07
Doesn't really matter if the discussion is in the issue or here -- they're both relevant places.

I'm in agreement with not increasing complexity for a small speedup.

And to be clear, I don't think the speedup is actually in the `pushBack()` call, but rather in the fact that using pushBack permits reading *directly* from the file connection, rather than a `textConnection`. Evidently something about the `read.table(text=...)` is what was scaling poorly and slowing it down.