@dankelley 2016-04-07
You're a lot better with GH than I am. Must be your youth! I won't admit it's your superior intelligence :-)

I did some more tests on the 14MB file. Writing
```r
        data <- as.list(read.table(file, skip=iline-1, header=FALSE, col.names=col.names.inferred,
                                   comment.char="", colClasses="numeric"))
 ```
instead of
```r
        data <- as.list(read.table(file, skip=iline-1, header=FALSE, col.names=col.names.inferred)
 ```
speeds things from 1.82s to 1.57s. The speedup is almost entirely from the `colClasses`.  However, this 15 percent speedup is not very impressive, and I worry that some files might have e.g. flags that are character-based, so I do not suggest using this in the code.

From my perspective, Clark, you've fixed issue #915.