@richardsc 2016-04-07
> You're a lot better with GH than I am. Must be your youth! I won't admit it's your superior intelligence :-)

Zing! Ouch!